make scale() double

+++ src/phaHX711.cpp
@@ -55,7 +55,7 @@ double phaHX711::factor(double weight) {
 }
 
 // output
-long phaHX711::scale() {
+double phaHX711::scale() {
   return (read()/100 - tareValue) / factorValue;
 }
